refactor: user 관련 클래스 별도 패키지로 분리
