+++ src/main/java/domain/card/Deck.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public interface Deck {
     Card draw();
