+++ src/test/java/domain/DeterminedDeck.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.Deck;
 import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.EmptyStackException;
