+++ src/test/java/domain/HandTest.java
@@ -2,6 +2,10 @@ package domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import domain.card.Card;
+import domain.card.Hand;
+import domain.card.Rank;
+import domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
