+++ src/test/java/domain/user/PlayerTest.java
@@ -1,9 +1,9 @@
 package domain.user;
 
-import domain.Card;
-import domain.Hand;
-import domain.Rank;
-import domain.Suit;
+import domain.card.Card;
+import domain.card.Hand;
+import domain.card.Rank;
+import domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
